Fix calculation of measured omega values
The `ome_eval` comes from simulated rotation series angles, and may not
be the best thing to use as a reference point when computing measured
omega values. We've had a hard time figuring out why it is being used.
The measurement of a spot's omega value should ideally not depend on
the current instrument calibration.

Instead, it seems better to use the code that is commented out
immediately above it, which uses omega edges from the frame data as
a reference point when computing measured omega values.

We have some new spot detection code (not in this repository) that also
measures omega values, and our new spot detection code is in much
greater agreement with `pull_spots()` if we implement this change.
That provides further evidence as well that this change produces more
accurate omega values for the spots.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>