Handle recursion in imageseries get_region()
Sometimes (as is the case for `OmegaImageSeries`),
`self._adapter` is itself an imageseries, rather than an imageseries
adapter.

To handle this recursion, use a forwarding function on the
`ImageSeries` base class to forward the call to its adapter (which,
if the adapter is itself an imageseries, will then forward the call
to its own adapter).

This fixes running [this example](https://github.com/HEXRD/examples/blob/master/NIST_ruby/multiruby_dexelas/include/mruby_config_composite_roi.yml).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>