Add option for closing file in HDF5 imageseries
Previously, the HDF5 file would always be closed automatically. However,
this presents a problem if there are multiple imageseries using the same
file handle, and one of them is deleted, because the file will be closed
and the other imageseries will become invalid.

This is particularly important for state files in the GUI, where there
are many images saved within a single HDF5 state file, and if a user removes
one detector, the file would be closed automatically and the other
image series would become invalid.

This PR adds the ability to prevent automatic closing of the file when
the imageseries is finished. `close_when_finished` is an optional
keyword argument that defaults to `True` if missing. It fixes the
issue with HDF5 state files.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>