Assume FIDDLE if not TARDIS or PXRDIP
The list of names of detectors was working for FIDDLE for the most
part. However, unlike TARDIS and PXRDIP, the FIDDLE instrument can
be projected to a cartesian view, and in the GUI, we use a fake
detector for the cartesian view which is named `dpanel`. We cannot
automatically determine the instrument based upon the name `dpanel`.
So in order for the Cartesian view to work, we need to just assume
the instrument is `FIDDLE` if the detector names don't match TARDIS
and PXRDIP.

We ought to come up with a more robust solution for determining
which `eHat_l` and `eta_shift` to use for different instruments.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>