Significantly improve fit-grains performance
Profiling showed that the list comprehension in `objFuncFitGrain`
was taking a significant amount of time. This is because that
function is called repeatedly during the least squares fit, and
repeatedly performing the list comprehension would take a substantial
amount of time.

Performing the list comprehension outside of `objFuncFitGrain` was
shown to provide a substantial speedup.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>