Memoize KDTree creation
Profiling indicated that creating the KDTree takes a substantial amount
of time. We can improve performance by a lot (even in the multiprocessing
case) by memoizing the KDTree.

We always increase the minimum memoization size to the number of detectors
so that we can always at least memoize all of the KDTrees for a single
instrument.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>