Remove spots assigned to more than one HKL
We have a very hard time running fit-grains correctly when there
are spots assigned to more than one HKL. For now, we should just
ignore them.

Many of the spots assigned to more than one HKL are overlapping
spots which were mistakenly identified as a single spot. When we
have better spot finding, we may be able to skip this step. However,
it is still essential that each peak only be assigned to one HKL.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>