Generalize physics package layers and distortion
This simplifies and generalizes the layers in a physics package, so that
more can be easily added and used.

It adds three new default layers to the HED physics package: ablator,
heatshield, and pusher.

It also generalizes the "sample layer distortion" to be just
"layer distortion". All of the sample layer distortion calculations can
actually be applied to any layer. The only difference is that the layer
standoff (distance to pinhole) will be different.

We can use this to apply appropriate two theta distortions for layers
such as the ablator, heatshield, pusher, or even window.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>