Add chemical formula to layer properties
For both detector coating layers and physics package layers, allow a
chemical formula to be provided.

This fixes an issue where you previously always had to provide the
chemical formula for the material name. Now, you can provide other names
such as `diamond` and the chemical formula can be provided separately
(for example, `C8`) to be used for absorption correction calculations.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>