Ensure lmfit parameters don't contain '-'
lmfit parameters cannot contain '-', or they produce an error.

Both the detector names and material names are user-defined, and
they can contain '-' if the user adds it. We need to make sure we
convert those to '_' in parameter names, though, so that lmfit will
accept them.

We were already doing this for detector names, and we just needed
to add this sanitization for material names as well.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>