Memoize frame cache load
This is done so that, if multiple imageseries are using the same file
(such as what happens for 32-subpanel Eiger), they can all share the same
sparse matrices, and the data only needs to be loaded in one time.

This saves a significant amount of RAM and disk loading time, especially
for 32-subpanel Eiger instruments, since the data only needs to be loaded
in one time and shared between the subpanels.

Fixes: #808

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>