Perform fancy indexing on csr_matrices
Rather than expanding the full csr_matrix into a numpy array, then
performing fancy indexing on that, do the reverse: perform the
fancy indexing on the csr_matrix, and then expand it into a numpy
array.

This has the potential to save huge amounts of computing time, since
we do not have to create the entire dense array before performing
the indexing.

For the multiruby test case, this did not show any change to the
performance. However, for a monolithic Eiger example from Josh Ward,
it showed a 5x speed improvement.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>