Add beam energy correction to grains
This allows a slope (in eV/mm), intercept (in mm), and axis ('x' or 'y')
to be specified for a beam energy correction according to grain positions.

This is important to take into consideration gradients in the beam energy
profile, such that some grains may "experience" a different beam energy
than others.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>