Use expression for phase fraction of last material
Use an expression to ensure the phase fraction of the last material
results in a total phase fraction of 1.

To do this, we had to pass all parameters to the minimizer. This
resulted in some other issues we had to fix as well, including a
need to reset the background parameters during initialization, and
making the phase classes use correct iterators.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>