Use np.load() within a context manager
The NpzFile class does not automatically close the file. This can lead
to issues especially when multiple applications are reading/writing
to the same file.

We noticed that if the GUI had an npz file open, and then in Python
scripts we overwrote the npz file, that it would not overwrite
correctly.

Using the context manager ensures that the file gets closed
automatically when we are done with it, and should hopefully
resolve this issue.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>