Make applying the panel buffer to image optional
This can be a time-consuming operation to perform repeatedly. When we
reuse the same images over and over again, we should apply the panel
buffer beforehand and skip this step when generating the polar view.

This results in yet another 40% speedup for generating a bunch of
high resolution polar images for 32-subpanel eiger.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>