Make memoized output numpy arrays readonly
When a memoized function returns a numpy array, it is possible that
the caller of the function may modify the numpy array. This
unfortunately also modifies the cached numpy array stored within
the memoize decorator, and thus affects future callers.

As such, make returned numpy arrays read-only so that we do not
run into these hard-to-track-down bugs.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>