Ensure areas are positive in auto powder picking
The areas can be negative if the beam vector is in the opposite
direction from which HEXRD would normally expect it (HEXRD normally
expects a negative Z beam vector).

The powder picking code was assuming the areas are positive. Take
the absolute value to ensure that is the case. This fixes automatic
point picking for positive beam vectors.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>