Update Linux packaging to build using AlmaLinux 8
Centos 7 reached its end of life in 2024. manylinux2014 also reached
its end of life at the same time.

AlmaLinux 8 is used for manylinux_2_28 support. This supports glibc
2.28 and later, which supports Linux OS's starting in August 2018.
AlmaLinux 8 will be supported with security updates until 2029.

These updates are necessary, not only because of the end of life
dates that have been reached, but also because conda must now be
ran on a system supporting glibc 2.28.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>