Allow known panel buffers to be named by string
This allows us to use the much simpler, easier to read, and much
smaller in disk space instrument yaml files, instead of the .hexrd files,
while specifying known panel buffers.

For example, a panel buffer can now be specified like this in the yaml
file:

```yaml
  buffer: 'chess-eiger-stream-v1'
```

This will automatically load and use the panel buffer when needed.

It can also be used for subpanels that have an `roi` specification, and
the region needed from the buffer will be extracted automatically using
the `roi`.

The npz files are very small (around 80 kb), so they don't add a ton of
space to the git history, and they should also never change.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>